requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
alembic>=1.13.1
mysql-connector-python>=8.2.0
pymysql>=1.1.0
pydantic>=2.6.0
pydantic[email]>=2.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
python-decouple>=3.8
apscheduler>=3.10.4
